[Application] Move tests to appropriate namespace to match main component structure.
